+++ libs/@rustymotors/binary/src/lib/BareCString.ts
