+++ libs/@rustymotors/binary/src/lib/BinaryMember.ts
