+++ libs/@rustymotors/binary/src/lib/Bytable.ts
