+++ libs/@rustymotors/binary/src/lib/BytableBase.ts
