+++ libs/@rustymotors/binary/src/lib/BytableBitField.ts
