+++ libs/@rustymotors/binary/src/lib/BytableByte.ts
