+++ libs/@rustymotors/binary/src/lib/BytableCString.ts
