+++ libs/@rustymotors/binary/src/lib/BytableChannelData.ts
