+++ libs/@rustymotors/binary/src/lib/BytableContainer.ts
