+++ libs/@rustymotors/binary/src/lib/BytableData.ts
