+++ libs/@rustymotors/binary/src/lib/BytableDword.ts
