+++ libs/@rustymotors/binary/src/lib/BytableHeader.ts
