+++ libs/@rustymotors/binary/src/lib/BytableMessage.ts
