+++ libs/@rustymotors/binary/src/lib/BytableServerMessage.ts
