+++ libs/@rustymotors/binary/src/lib/BytableWord.ts
