+++ libs/@rustymotors/binary/src/lib/NpsRiffInfo.ts
