+++ libs/@rustymotors/binary/src/lib/Serializer.ts
