+++ libs/@rustymotors/binary/src/lib/binary.ts
