+++ libs/@rustymotors/binary/src/lib/types.ts
