+++ libs/@rustymotors/logging/src/types.ts
