+++ package.json
@@ -43,6 +43,7 @@
         "@dotenvx/dotenvx": "^1.45.1",
         "@fastify/sensible": "^6.0.3",
         "@roarr/sentry": "^3.0.0",
+        "@rustymotors/binary": "file:libs/@rustymotors/binary",
         "@rustymotors/network": "file:libs/@rustymotors/network",
         "@rustymotors/parser": "file:libs/@rustymotors/parser",
         "@rustymotors/rooms": "file:libs/@rustymotors/rooms",
@@ -87,6 +88,7 @@
     "packageManager": "npm@11.6.0+sha512.77f3fb0dbbd881835d7bd1217deabdf7ed77fc4ec4f363df64fc3cb986404abf6c437661041080f5c5d225103508e7c9ea30cb2742b7e942675d05a10af2d7b9",
     "workspaces": [
         "libs/*",
+        "libs/@rustymotors/*",
         "packages/*",
         "src/*"
     ]
