+++ packages/gateway/package.json
@@ -20,7 +20,7 @@
 	"license": "AGPL-3.0",
 	"dependencies": {
 		"@fastify/sensible": "^6.0.3",
-		"@rustymotors/binary": "",
+		"@rustymotors/binary": "file:libs/@rustymotors/binary",
 		"@rustymotors/logging": "",
 		"@sentry/cli": "^3.4.1",
 		"@sentry/node": "^10.52.0",
