+++ packages/shard/package.json
@@ -26,6 +26,7 @@
 		"vitest": "^4.1.5"
 	},
 	"dependencies": {
+		"@rustymotors/binary": "file:libs/@rustymotors/binary",
 		"@sentry/cli": "^3.4.1",
 		"@sentry/node": "^10.52.0",
 		"@sentry/profiling-node": "^10.52.0",
