+++ vitest.config.ts
@@ -7,8 +7,7 @@ import { existsSync } from "node:fs";
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 
-// Get project root (two levels up from packages/gateway)
-const projectRoot = resolve(__dirname, "../..");
+const projectRoot = __dirname;
 const envPath = resolve(projectRoot, ".env");
 
 // Only include --env-file if .env exists (for CI compatibility)
@@ -19,10 +18,10 @@ if (existsSync(envPath)) {
 
 export default defineConfig({
 	test: {
-		setupFiles: ["./vitest.setup.ts"],
+		setupFiles: ["./vitest.setup.mjs"],
 		globals: true,
 		environment: "node",
-		// Exclude session replay tests from default test run
+		// Exclude session replay tests and stale build output from default test run
 		exclude: [
 			"**/node_modules/**",
 			"**/dist/**",
@@ -32,23 +31,21 @@ export default defineConfig({
 			"**/session/integration.example.test.ts",
 			"**/session/SessionRecorder.test.ts",
 		],
-		// Enable Node's built-in .env file support and legacy crypto
-		poolOptions: {
-			forks: {
-				// Use Node's built-in --env-file support (Node 20.6+)
-				// Only include if .env file exists (for CI compatibility)
-				execArgv,
-			},
-		},
 		server: {
 			deps: {
 				inline: ["rusty-motors-protocol", "@rustymotors/binary"],
 			},
 		},
 	},
+	// Vitest 4: poolOptions moved out from under `test`.
+	poolOptions: {
+		forks: {
+			// Use Node's built-in --env-file support (Node 20.6+) and legacy crypto.
+			execArgv,
+		},
+	},
 	resolve: {
 		alias: {
-			"@": resolve(__dirname, "./src"),
 			"rusty-motors-protocol": resolve(projectRoot, "packages/protocol"),
 			"@rustymotors/binary": resolve(projectRoot, "libs/@rustymotors/binary"),
 		},
